Disable SSL health checks by default
This is a temporary change to allow dataplaneapi to override the
client's request for health checks over SSL. While it could definitely
be done more elegantly, eg. via a configuration option, or by
implementing a regular expression that allowed more flexibility at
runtime, this is a temporary patch to validate that health checks work
as expected.